+++ packages/api/src/Domain/Response/Subscription/SubscriptionInviteListResponse.ts
@@ -1,7 +1,9 @@
+import { Either } from '@standardnotes/common'
+
 import { HttpErrorResponseBody } from '../../Http/HttpErrorResponseBody'
 import { HttpResponse } from '../../Http/HttpResponse'
 import { SubscriptionInviteListResponseBody } from './SubscriptionInviteListResponseBody'
 
 export interface SubscriptionInviteListResponse extends HttpResponse {
-  data: SubscriptionInviteListResponseBody | HttpErrorResponseBody
+  data: Either<SubscriptionInviteListResponseBody, HttpErrorResponseBody>
 }