+++ packages/api/src/Domain/Server/index.ts
@@ -1,3 +1,4 @@
-export * from './User/Paths'
+export * from './Subscription/SubscriptionServer'
+export * from './Subscription/SubscriptionServerInterface'
 export * from './User/UserServer'
 export * from './User/UserServerInterface'