+++ packages/features/src/Domain/Feature/FeatureIdentifier.ts
@@ -17,6 +17,7 @@ export enum FeatureIdentifier {
   SmartFilters = 'org.standardnotes.smart-filters',
   TagNesting = 'org.standardnotes.tag-nesting',
   TwoFactorAuth = 'org.standardnotes.two-factor-auth',
+  SubscriptionSharing = 'org.standardnotes.subscription-sharing',
 
   AutobiographyTheme = 'org.standardnotes.theme-autobiography',
   DynamicTheme = 'org.standardnotes.theme-dynamic',