+++ packages/models/src/Domain/Device/Platform.ts
@@ -1,10 +1,3 @@
-export enum Environment {
-  Web = 1,
-  Desktop = 2,
-  Mobile = 3,
-  NativeMobileWeb = 4,
-}
-
 export enum Platform {
   Ios = 1,
   Android = 2,