+++ packages/services/src/Domain/Subscription/SubscriptionClientInterface.ts
@@ -1,0 +1,8 @@
+import { Uuid } from '@standardnotes/common'
+import { Invitation } from '@standardnotes/models'
+
+export interface SubscriptionClientInterface {
+  listSubscriptionInvitations(): Promise<Invitation[]>
+  inviteToSubscription(inviteeEmail: string): Promise<boolean>
+  cancelInvitation(inviteUuid: Uuid): Promise<boolean>
+}