+++ packages/snjs/lib/Services/Api/HttpService.ts
@@ -2,7 +2,8 @@ import { API_MESSAGE_RATE_LIMITED, UNKNOWN_ERROR } from './Messages'
 import { HttpResponse, StatusCode } from '@standardnotes/responses'
 import { isString } from '@standardnotes/utils'
 import { SnjsVersion } from '@Lib/Version'
-import { AbstractService, InternalEventBusInterface, Environment } from '@standardnotes/services'
+import { AbstractService, InternalEventBusInterface } from '@standardnotes/services'
+import { Environment } from '@standardnotes/models'
 
 export enum HttpVerb {
   Get = 'GET',