+++ packages/desktop/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.23.117](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.116...@standardnotes/desktop@3.23.117) (2022-09-13)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
+## [3.23.116](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.115...@standardnotes/desktop@3.23.116) (2022-09-13)
+
+**Note:** Version bump only for package @standardnotes/desktop
+
 ## [3.23.115](https://github.com/standardnotes/app/compare/@standardnotes/desktop@3.23.114...@standardnotes/desktop@3.23.115) (2022-09-12)
 
 **Note:** Version bump only for package @standardnotes/desktop